% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model_Unscaled_nuts_plant}
\alias{create_model_Unscaled_nuts_plant}
\title{Initialize an ATN model, following Schneider et al. 2016, Nature Communication}
\usage{
create_model_Unscaled_nuts_plant(nb_b, nb_n = 2, BM)
}
\arguments{
\item{nb_b}{integer, number of basal species.}

\item{nb_n}{integer, number of nutrients.}

\item{BM}{float vector, body mass of basal species.}

\item{faci}{facilitation matrix of primary producers ranging from 0 to 1, diagnol set to 0.}
}
\value{
An object of class \emph{ATN (Rcpp_parameters_prefs)}.
}
\description{
Initialize an ATN model, following Schneider et al. 2016, Nature Communication
}
\details{
A model is defined by the total number of species
(\emph{nb_s}), the number of basal species (\emph{nb_b}),
the number of nutrients (\emph{nb_n}), the body masses
 (\emph{BM}) of basal species, this model doesn't have consumers. 
Nutrients are not counted as species.
}
