% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_extinct.R
\name{.filter_extinct}
\alias{.filter_extinct}
\title{Filter Extinct Species}
\usage{
.filter_extinct(df, model)
}
\arguments{
\item{df}{deSolve matrix as returned from lsoda_wrapper().}

\item{model}{ATNr model object, from which extinction threshold is extracted.}
}
\value{
df with values below th set to zero.
}
\description{
Filter Extinct Species
}
\details{
Set to zero species biomass that are below the extinction threshold.
}
