% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_species.R
\name{.remove_Binzer}
\alias{.remove_Binzer}
\title{Internal function to remove species from a Binzer model}
\usage{
.remove_Binzer(species, model)
}
\description{
Internal function to remove species from a Binzer model
}
\keyword{internal}
