% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_species.R
\name{.remove_Scaled}
\alias{.remove_Scaled}
\title{Internal function to remove species from a Scaled model}
\usage{
.remove_Scaled(species, model)
}
\description{
Internal function to remove species from a Scaled model
}
\keyword{internal}
