% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_species.R
\name{.remove_Unscaled}
\alias{.remove_Unscaled}
\title{Internal function to remove species from a Unscaled model}
\usage{
.remove_Unscaled(species, model)
}
\description{
Internal function to remove species from a Unscaled model
}
\keyword{internal}
