% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_species.R
\name{.remove_Unscaled_nuts}
\alias{.remove_Unscaled_nuts}
\title{Internal function to remove species from a Unscaled_nuts model}
\usage{
.remove_Unscaled_nuts(species, model, nuts)
}
\description{
Internal function to remove species from a Unscaled_nuts model
}
\keyword{internal}
