% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_initialisations.R
\name{initialise_default_Delmas_2016}
\alias{initialise_default_Delmas_2016}
\title{Default parameters for the scaled version of ATN as in Delmas et al.
  2016}
\usage{
initialise_default_Delmas_2016(model)
}
\arguments{
\item{model}{an object of class \emph{ATN (Delmas), Delmas}.}
}
\value{
An object of class \emph{ATN (Delmas)} with default
  parameters as in Delmas et al. (2017).
}
\description{
Initialise the default parametrisation for the scaled version of
  the ATN model as in Delams et al. (2016).
}
\references{
Delmas, E., Brose, U., Gravel, D., Stouffer, D.B. and Poisot, T.
  (2017), Simulations of biomass dynamics in community food webs. Methods
  Ecol Evol, 8: 881-886. https://doi.org/10.1111/2041-210X.12713
}
