% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_initialisations.R
\name{initialise_default_Unscaled_faci}
\alias{initialise_default_Unscaled_faci}
\title{Default model parameters as in Schneider et al. 2016 with option to manipulate facilitation}
\usage{
initialise_default_Unscaled_faci(model, L.mat, temperature = 20)
}
\arguments{
\item{model}{an object of class \emph{ATN (Rcpp_Unscaled_nuts}.}

\item{L.mat}{numeric matrix, probability of a consumer to attack and capture an encountered resource. See \code{\link{create_Lmatrix}}.}

\item{temperature}{numeric, ambient temperature of the ecosystem in Celsius.}
}
\value{
An object of class \emph{ATN (Rcpp_Unscaled_nuts_faci)} with default
  parameters as in Schneider et al. (2016).
}
\description{
Initialise the default parametrisation for the model for
  Schneider et al. (2016).
}
