% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_initialisations.R
\name{initialise_default_Unscaled_nuts}
\alias{initialise_default_Unscaled_nuts}
\title{Default model parameters as in Schneider et al. 2016}
\usage{
initialise_default_Unscaled_nuts(model, L.mat, temperature = 20)
}
\arguments{
\item{model}{an object of class \emph{ATN (Rcpp_Unscaled_nuts}.}

\item{L.mat}{numeric matrix, probability of a consumer to attack and capture an encountered resource. See \code{\link{create_Lmatrix}}.}

\item{temperature}{numeric, ambient temperature of the ecosystem in Celsius.}
}
\value{
An object of class \emph{ATN (Rcpp_Unscaled_nuts)} with default
  parameters as in Schneider et al. (2016).
}
\description{
Initialise the default parametrisation for the model for
  Schneider et al. (2016).
}
\references{
Schneider, F. D., Brose, U., Rall, B. C., & Guill, C. (2016).
  Animal diversity and ecosystem functioning in dynamic food webs. Nature
  Communications, 7(1), 1-8.
}
