% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_initialisations.R
\name{initialise_default_Unscaled_plant}
\alias{initialise_default_Unscaled_plant}
\title{Default model parameters as in Schneider et al. 2016 but without consumers with option to manipulate facilitation}
\usage{
initialise_default_Unscaled_plant(model, temperature = 20)
}
\arguments{
\item{model}{an object of class \emph{ATN (Rcpp_Unscaled_nuts_plant}.}

\item{temperature}{numeric, ambient temperature of the ecosystem in Celsius.}
}
\value{
An object of class \emph{ATN (Rcpp_Unscaled_nuts_plant)} with default
  parameters as in Schneider et al. (2016).
}
\description{
Initialise the default parametrisation for the model for
  Schneider et al. (2016).
}
