% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_species.R
\name{remove.Delmas}
\alias{remove.Delmas}
\title{Internal function to remove species from a Delmas model}
\usage{
remove.Delmas(species, model)
}
\description{
Internal function to remove species from a Delmas model
}
\keyword{internal}
