% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_species.R
\name{remove.Scheinder}
\alias{remove.Scheinder}
\title{Internal function to remove species from a Scheinder model}
\usage{
remove.Scheinder(species, model, nuts)
}
\description{
Internal function to remove species from a Scheinder model
}
\keyword{internal}
