% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_species.R
\name{remove_species}
\alias{remove_species}
\title{Function to remove species from a model class}
\usage{
remove_species(species, model, nuts = NULL)
}
\arguments{
\item{species}{integer vector, the indices of species to remove.}

\item{model}{model object}

\item{nuts}{integer vector, the indices of nutrients to remove. Parameter
specific to the Unscaled_nuts model.}
}
\value{
A model object where the data structure has bee updated to remove the
  species in parameters.
}
\description{
Function to remove species from a model class
}
