% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_checks}
\alias{run_checks}
\title{Run checks on model parameters}
\usage{
run_checks(model, verbose = TRUE)
}
\arguments{
\item{model}{a model object, created from \code{\link{create_model_schneider}}) or
\code{\link{create_model_delmas}})}
}
\description{
Check if the dimensions of vectors and matrices used in the model are correct.
If any dimension is not correct, an error message is returned.
}
\examples{
library(ATN) # is an example truly needed here?
}
