% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_solvers.R
\name{sundial_wrapper}
\alias{sundial_wrapper}
\title{Wrapper for sundial}
\usage{
sundial_wrapper(t, y, model)
}
\arguments{
\item{t}{vector of times.}

\item{y}{vector of biomasses.}

\item{model}{object of class \emph{ATN (Rcpp_parameters_prefs)}.}
}
\value{
A matrix for the ODE solution with species as columns and
times as rows.
}
\description{
This is a wrapper to call \code{\link{cvode}} from
\emph{sundialr} and solve the ODE.
Package \code{\link{sundialr}} needs to be installed to run
this wrapper.
}
\examples{
library(ATN)
masses <- runif(50, 10, 100) #body mass of species
L <- create_Lmatrix(masses, 10, Ropt = 50)
L[L > 0] <- 1
mod <- create_model_schneider(50, 10, 3, masses, L)
mod <- initialise_default_Schneider_2016(mod, L)
biomasses <- masses ^ -0.75 * 10 ^ 4 #biomasses of species
biomasses <- append(runif(3, 20, 30), biomasses)
times <- seq(0, 100, 1)
#' sol <- sundial_wrapper(times, biomasses, mod)
t <- times
y <- biomasses
sundialr::cvode(
time_vector = 0.0, #time vectors
IC = y, #initial conditions
input_function = function(t, y, p) mod$ODE(y, t), #anonymous function to reorder input
Parameters = c(0, 0) #this does nothing, but is necessary for compatibility
)
}
\keyword{internal}
