% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_level.R
\name{TroLev}
\alias{TroLev}
\title{Calculate trophic level of species}
\usage{
TroLev(fw)
}
\arguments{
\item{fw}{numeric matrix, the matrix of the food web.}
}
\value{
A numeric vector of species' trophic level.
}
\description{
Calculate trophic level of species
}
\examples{
library(ATNr)
set.seed(123)
# create a food web from the niche model with 35 species and connectance of 0.1
fw <- create_niche_model(20, 0.1)
TL = TroLev(fw)
 
}
