% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Unscaled_nuts_loops}
\alias{Unscaled_nuts_loops}
\title{Store parameters and functions associated to the unscaled version of ATN}
\description{
To not use. For testing purpose only. Please use Rcpp_Unscaled_nuts instead.
}
