% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Unscaled_nuts_plant}
\alias{Unscaled_nuts_plant}
\title{Store parameters and functions associated to the unscaled version of ATN including nutrient dynamics}
\description{
Type the name of the class to see its methods
}
\section{Fields}{

\describe{
\item{\code{nb_b}}{Number of basal species (only plants)}

\item{\code{nb_n}}{Number of nutrient pool}

\item{\code{X}}{vector of metabolic rates (length = number of species)}

\item{\code{K}}{matrix of plant nutrient efficiencies (dim = number of nutrients * number of plants)}

\item{\code{V}}{matrix of plant relative nutrient content (dim = number of nutrients * number of plants)}

\item{\code{S}}{Vector of maximum nutrient concentration (length = number of plants)}

\item{\code{r}}{Vector of maximum growth rate of plant species (length = number of plant species)}

\item{\code{BM}}{Vector of body masses (length = number of species)}

\item{\code{dB}}{Vector of local derivatives (length = number of species)}

\item{\code{faci}}{Matrix of plant-plant facilitation schemes}

\item{\code{ext}}{Extinction threshold for species}

\item{\code{theta}}{the exponent of neighbour tree biomass to affect the realized growth}

\item{\code{ODE}}{Calculate the derivatives for the scaled version of the ATN model \itemize{
\item Parameter: bioms -  Local species biomasses
\item Parameter: t - Integration time point
\item Returns a vector of growth rate for each species at time t
}}
}}

