% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_checks}
\alias{run_checks}
\title{Run checks on model parameters}
\usage{
run_checks(model, verbose = TRUE)
}
\arguments{
\item{model}{a model object.}

\item{verbose}{Boolean, whether a message should be printed when all checks were successful}
}
\description{
Check if the dimensions of vectors and matrices used in the model are correct.
If any dimension is not correct, an error message is returned.
}
