% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_level.R
\name{TroLev}
\alias{TroLev}
\title{Calculate trophic level of species}
\usage{
TroLev(fw)
}
\arguments{
\item{fw}{numeric matrix, the matrix of the food web.}
}
\value{
A numeric vector of species' trophic level.
}
\description{
Calculate trophic level of species
}
